// packages/core/src/enums.ts
// Business enumerations
export enum Industry {
  TECHNOLOGY = 'technology',
  HEALTHCARE = 'healthcare',
  FINANCE = 'finance',
  MANUFACTURING = 'manufacturing',
  RETAIL = 'retail',
  EDUCATION = 'education',
  REAL_ESTATE = 'real_estate',
  CONSTRUCTION = 'construction',
  TRANSPORTATION = 'transportation',
  HOSPITALITY = 'hospitality',
  AGRICULTURE = 'agriculture',
  ENERGY = 'energy',
  MEDIA = 'media',
  CONSULTING = 'consulting',
  LEGAL = 'legal',
  OTHER = 'other'
}

export enum BusinessType {
  CORPORATION = 'corporation',
  LLC = 'llc',
  PARTNERSHIP = 'partnership',
  SOLE_PROPRIETORSHIP = 'sole_proprietorship',
  NON_PROFIT = 'non_profit',
  COOPERATIVE = 'cooperative'
}

export enum BusinessMode {
  ONLINE = 'online',
  OFFLINE = 'offline',
  HYBRID = 'hybrid'
}

export enum Ownership {
  PUBLIC = 'public',
  PRIVATE = 'private',
  GOVERNMENT = 'government',
  FRANCHISE = 'franchise'
}

export enum RevenueBand {
  UNDER_1M = 'under_1m',
  RANGE_1M_5M = 'range_1m_5m',
  RANGE_5M_10M = 'range_5m_10m',
  RANGE_10M_50M = 'range_10m_50m',
  RANGE_50M_100M = 'range_50m_100m',
  RANGE_100M_500M = 'range_100m_500m',
  RANGE_500M_1B = 'range_500m_1b',
  OVER_1B = 'over_1b'
}

export enum EmployeeBand {
  UNDER_10 = 'under_10',
  RANGE_10_50 = 'range_10_50',
  RANGE_50_100 = 'range_50_100',
  RANGE_100_500 = 'range_100_500',
  RANGE_500_1000 = 'range_500_1000',
  RANGE_1000_5000 = 'range_1000_5000',
  RANGE_5000_10000 = 'range_5000_10000',
  OVER_10000 = 'over_10000'
}